refactor: split auth token and mfa primitives
